+++ bfvmm/include/vcpu/arch/intel_x64/vcpu.h
@@ -20,11 +20,11 @@
 #define VCPU_INTEL_X64_H
 
 #include <vcpu/vcpu.h>
-#include <hve/arch/intel_x64/vmxon/vmxon.h>
-#include <hve/arch/intel_x64/vmcs/vmcs.h>
-#include <hve/arch/intel_x64/vmcs/vmcs_state_vmm.h>
-#include <hve/arch/intel_x64/vmcs/vmcs_state_hvm.h>
-#include <hve/arch/intel_x64/exit_handler/exit_handler.h>
+#include <bfvmm/hve/arch/intel_x64/vmxon/vmxon.h>
+#include <bfvmm/hve/arch/intel_x64/vmcs/vmcs.h>
+#include <bfvmm/hve/arch/intel_x64/vmcs/vmcs_state_vmm.h>
+#include <bfvmm/hve/arch/intel_x64/vmcs/vmcs_state_hvm.h>
+#include <bfvmm/hve/arch/intel_x64/exit_handler/exit_handler.h>
 
 // -----------------------------------------------------------------------------
 // Exports
