+++ bfvmm/src/hve/arch/intel_x64/vmxon/vmxon.cpp
@@ -19,7 +19,7 @@
 #include <bfgsl.h>
 #include <bfdebug.h>
 
-#include <hve/arch/intel_x64/vmxon/vmxon.h>
+#include <bfvmm/hve/arch/intel_x64/vmxon/vmxon.h>
 #include <memory_manager/memory_manager_x64.h>
 
 #include <intrinsics.h>
