+++ bfvmm/tests/CMakeLists.txt
@@ -21,7 +21,7 @@ project(bfvmm_test C CXX)
 
 include(${SOURCE_CMAKE_DIR}/project.cmake)
 init_project(
-    INCLUDES ${CMAKE_CURRENT_LIST_DIR}/../include
+    INCLUDES ${CMAKE_CURRENT_LIST_DIR}/../include ${CMAKE_CURRENT_LIST_DIR}/../../include
 )
 
 add_subdirectory(../src/debug ${CMAKE_CURRENT_BINARY_DIR}/src/debug)
