+++ include/bfvmm/hve/arch/intel_x64/exit_handler/exit_handler.h
@@ -23,7 +23,7 @@
 
 #include <intrinsics.h>
 #include <memory_manager/map_ptr_x64.h>
-#include <hve/arch/intel_x64/vmcs/vmcs.h>
+#include <bfvmm/hve/arch/intel_x64/vmcs/vmcs.h>
 
 #include <bfjson.h>
 #include <bfvmcallinterface.h>
