+++ include/bfvmm/hve/arch/intel_x64/vmcs/vmcs.h
@@ -22,8 +22,8 @@
 #include <bfdelegate.h>
 
 #include <intrinsics.h>
-#include <hve/arch/intel_x64/state_save.h>
-#include <hve/arch/intel_x64/vmcs/vmcs_state.h>
+#include <bfvmm/hve/arch/intel_x64/state_save.h>
+#include <bfvmm/hve/arch/intel_x64/vmcs/vmcs_state.h>
 
 // -----------------------------------------------------------------------------
 // Exports
