+++ include/bfvmm/hve/arch/intel_x64/vmcs/vmcs_check_host.h
