+++ include/bfvmm/hve/arch/intel_x64/vmcs/vmcs_promote.h
@@ -16,23 +16,20 @@
 // License along with this library; if not, write to the Free Software
 // Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA
 
-#ifndef EXIT_HANDLER_INTEL_X64_SUPPORT_H
-#define EXIT_HANDLER_INTEL_X64_SUPPORT_H
+#ifndef VMCS_INTEL_X64_PROMOTE_H
+#define VMCS_INTEL_X64_PROMOTE_H
 
-#include <hve/arch/intel_x64/exit_handler/exit_handler.h>
+#include <cstdint>
+#include <bfvmm/hve/arch/intel_x64/state_save.h>
 
-/// Exit Handler Entry
+/// Promote VMCS
 ///
-/// This is the starting point of the VMM. It is written in pure assembly
-/// in order to ensure the state of the guest is handled properly. This
-/// code saves / restores the guest's CPU state, and then hands control off
-/// to the "C" portion of the code to continue execution, and begin the
-/// process of handling the VM exit.
+/// Promote the guest described by this VMCS to VMX-root operation. In
+/// the process of doing this, the VMM's state is lost.
 ///
-/// @expects none
-/// @ensures none
+/// @note this function does not return
 ///
-extern "C" EXPORT_HVE void exit_handler_entry(
-    void) noexcept;
+extern "C" void vmcs_promote(
+    state_save_intel_x64 *state_save, const void *guest_gdt) noexcept;
 
 #endif
