+++ include/bfvmm/hve/arch/intel_x64/vmcs/vmcs_state.h
@@ -22,8 +22,8 @@
 #include <bftypes.h>
 
 #include <intrinsics.h>
-#include <hve/arch/x64/gdt.h>
-#include <hve/arch/x64/idt.h>
+#include <bfvmm/hve/arch/x64/gdt.h>
+#include <bfvmm/hve/arch/x64/idt.h>
 
 // -----------------------------------------------------------------------------
 // Exports
